KP-5-boids/requirements.txt
ffmpeg-python      0.2.0
numba              0.56.4
numpy              1.21.0
pip                22.3.1
pyglet             1.5.23
vispy              0.10.0

